+++ src/listeners.js
@@ -1,3 +1,6 @@
+import { actionOnSymbol, thunkOnSymbol } from './constants';
+import { get } from './lib';
+
 export function createListenerMiddleware(references) {
   return () => next => action => {
     const result = next(action);
@@ -20,4 +23,45 @@ export function createListenerMiddleware(references) {
     }
     return result;
   };
+}
+
+export function bindListenerDefinitions(
+  listenerDefinitions,
+  actionCreators,
+  actionCreatorDict,
+  listenerActionMap,
+) {
+  listenerDefinitions.forEach(listenerActionOrThunk => {
+    const listenerMeta =
+      listenerActionOrThunk[actionOnSymbol] ||
+      listenerActionOrThunk[thunkOnSymbol];
+
+    const targets = listenerMeta.targetResolver(
+      get(listenerMeta.parent, actionCreators),
+      actionCreators,
+    );
+    const targetTypes = (Array.isArray(targets) ? targets : [targets]).reduce(
+      (acc, target) => {
+        if (
+          typeof target === 'function' &&
+          target.type &&
+          actionCreatorDict[target.type]
+        ) {
+          acc.push(target.type);
+        } else if (typeof target === 'string') {
+          acc.push(target);
+        }
+        return acc;
+      },
+      [],
+    );
+
+    listenerMeta.resolvedTargets = targetTypes;
+
+    targetTypes.forEach(targetType => {
+      const listenerReg = listenerActionMap[targetType] || [];
+      listenerReg.push(actionCreatorDict[listenerMeta.type]);
+      listenerActionMap[targetType] = listenerReg;
+    });
+  });
 }
